re-organize all files in folders, simplify imports
