+++ felupe/__init__.py
@@ -7,12 +7,51 @@ from . import element
 from . import tools
 from . import constitution
 from . import solve
+from . import region
 
-from ._region import Region
-from ._form import IntegralForm, IntegralFormMixed, IntegralFormAxisymmetric
-from ._field import Field, FieldAxisymmetric, FieldMixed
-from .dof import Boundary
-
+from .region import Region
+from ._form import (
+    IntegralForm,
+    IntegralFormMixed,
+    IntegralFormAxisymmetric,
+)
+from ._field import (
+    Field,
+    FieldAxisymmetric,
+    FieldMixed,
+)
+from .dof import Boundary, MultiPointConstraint, MultiPointContact
+from .element import (
+    Line,
+    #
+    Quad,
+    ConstantQuad,
+    #
+    ArbitraryOrderLagrange as ArbitraryOrderLagrangeElement,
+    #
+    Hexahedron,
+    ConstantHexahedron,
+    QuadraticHexahedron,
+    TriQuadraticHexahedron,
+    #
+    Triangle,
+    TriangleMINI,
+    QuadraticTriangle,
+    #
+    Tetra,
+    TetraMINI,
+    QuadraticTetra,
+)
+from .quadrature import (
+    GaussLegendre,
+    Triangle as TriangleQuadrature,
+    Tetrahedron as TetrahedronQuadrature,
+)
+from .mesh import (
+    Mesh,
+    Rectangle,
+    Cube,
+)
 
 __all__ = [
     "__version__",
