+++ felupe/quadrature/__init__.py
@@ -1,0 +1,4 @@
+from ._base import Scheme
+from ._triangle import Triangle
+from ._tetra import Tetrahedron
+from ._gausslegendre import GaussLegendre
