+++ felupe/solve/__init__.py
@@ -1,0 +1,4 @@
+from ._solve import (
+    partition,
+    solve,
+)
