+++ felupe/tools/_field.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+"""
+ _______  _______  ___      __   __  _______  _______ 
+|       ||       ||   |    |  | |  ||       ||       |
+|    ___||    ___||   |    |  | |  ||    _  ||    ___|
+|   |___ |   |___ |   |    |  |_|  ||   |_| ||   |___ 
+|    ___||    ___||   |___ |       ||    ___||    ___|
+|   |    |   |___ |       ||       ||   |    |   |___ 
+|___|    |_______||_______||_______||___|    |_______|
+
+This file is part of felupe.
+
+Felupe is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Felupe is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with Felupe.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+
+from copy import deepcopy
+
+
+def update(y, dx, inplace=True):
+    "Update field values."
+
+    if inplace:
+        x = y
+    else:
+        x = deepcopy(y)
+
+    if isinstance(x, tuple) or isinstance(x, list):
+        for field, dfield in zip(x, dx):
+            field += dfield
+    elif "mixed" in str(type(x)):
+        for field, dfield in zip(x.fields, dx):
+            field += dfield
+    else:
+        x += dx
+
+    return x
