+++ felupe/tools/_solve.py
@@ -1,0 +1,57 @@
+# -*- coding: utf-8 -*-
+"""
+ _______  _______  ___      __   __  _______  _______ 
+|       ||       ||   |    |  | |  ||       ||       |
+|    ___||    ___||   |    |  | |  ||    _  ||    ___|
+|   |___ |   |___ |   |    |  |_|  ||   |_| ||   |___ 
+|    ___||    ___||   |___ |       ||    ___||    ___|
+|   |    |   |___ |       ||       ||   |    |   |___ 
+|___|    |_______||_______||_______||___|    |_______|
+
+This file is part of felupe.
+
+Felupe is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Felupe is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with Felupe.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+
+import numpy as np
+
+from .. import solve as solvetools
+
+
+def solve(K, f, field, dof0, dof1, ext, unstack=None):
+    "Solve linear equation system K dx = b"
+
+    if isinstance(field, tuple) or isinstance(field, list):
+        return _solve_mixed(K, f, field, dof0, dof1, ext, unstack)
+
+    elif "mixed" in str(type(field)):
+        return _solve_mixed(K, f, field, dof0, dof1, ext, unstack)
+
+    else:
+        return _solve_single(K, f, field, dof0, dof1, ext)
+
+
+def _solve_single(K, f, field, dof0, dof1, ext):
+    "Solve linear equation system K dx = b"
+    system = solvetools.partition(field, K, dof1, dof0, -f)
+    dx = solvetools.solve(*system, ext)
+    return dx
+
+
+def _solve_mixed(K, f, field, dof0, dof1, ext, unstack):
+    "Solve linear equation system K dx = b"
+    system = solvetools.partition(field, K, dof1, dof0, -f)
+    dfields = np.split(solvetools.solve(*system, ext), unstack)
+    return dfields
